+++ src/replicon_core.rs
@@ -1,4 +1,4 @@
-pub mod not_replicate;
+pub mod dont_replicate;
 pub mod replication_rules;
 pub mod replicon_tick;